Tweaks for initialization
Rename the `yes` param for init.sh to `auto`
Remove ~/.init with init.sh instead of .bashrc
Don't prompt for lastfm info when `auto` is passed to init.sh
Check for security restore script, prompt user to run it
Prompt user to reboot instead of doing so automatically

+++ .setup/init.sh
@@ -6,25 +6,26 @@
 #
 # This script is destructive and you probably shouldn't run it
 
-set -e
+set -ea
 echo ".dotfiles init"
-if [[ "$@" != "yes" ]]; then
+if [[ "$@" != "auto" ]]; then
    read -p "This script will overwrite existing files. Are you sure you want to run it? (Yes)"
    if [[ ! $REPLY =~ ^Yes$ ]]; then
       exit 1
    fi
+else
+   export LASTFM_USERNAME="squigglezworth"
+   export LASTFM_PASSWORD="foo"
 fi
 
 # Import colors, exporting for future use
-set -a
 [ -f $HOME/dotfiles/bash/colors ] && source $HOME/dotfiles/bash/colors
 [ -f $HOME/dotfiles/bash/scripts/color-utils.sh ] && source $HOME/dotfiles/bash/scripts/color-utils.sh
 
 # Gather user input, exporting for future use
-read -ep "Enter last.fm username: " LASTFM_USERNAME
-read -esp "Enter last.fm password: " LASTFM_PASSWORD; echo;
-# LASTFM_USERNAME="squigglezworth"
-# LASTFM_PASSWORD="test"
+[ ! LASTFM_USERNAME ] && read -ep "Enter last.fm username: " LASTFM_USERNAME
+[ ! LASTFM_PASSWORD ] && read -esp "Enter last.fm password: " LASTFM_PASSWORD; echo;
+
 sudo echo -n
 
 echoDone() { echo -e "${BGreen}Done!${Color_Off}" ; }
@@ -45,11 +46,26 @@ echo -e "\nInstalling:"
 echo -e "\nConfiguring:"
 .setup/config.sh
 
-echo -n "${BRed}Rebooting ... "
-secs=5
-while [ $secs -gt 0 ]; do
-   echo -ne "${BRed}Rebooting ... $secs \033[0K\r${Color_Off}"
-   sleep 1
-   ((secs--))
-done
-sudo reboot
+echo -e $(rainbow "All done! \o/")
+
+echo -e "Don't forget to restore sensitive files such as SSH keys!"
+if [ -f /mnt/usb/.restore.sh ]; then
+   read -p "Restore script found! Would you like to run it?"
+   if [[ $REPLY =~ [Yy] ]]; then
+      /mnt/usb/.restore.sh
+   fi
+fi
+
+rm ~/.init
+
+read -p "Would you like to reboot? (Y/y)"
+if [[ $REPLY =~ [Yy] ]]; then
+   echo -n "${BRed}Rebooting ... "
+   secs=5
+   while [ $secs -gt 0 ]; do
+      echo -ne "${BRed}Rebooting ... $secs \033[0K\r${Color_Off}"
+      sleep 1
+      ((secs--))
+   done
+   sudo reboot
+fi
